Rename usersFriendsRoute to userFriendsRoute
